requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
APScheduler==3.10.4
async-timeout==4.0.3
attrs==23.2.0
discord.py==2.3.2
frozenlist==1.4.1
idna==3.6
multidict==6.0.5
pytz==2024.1
six==1.16.0
tzlocal==5.2
yarl==1.9.4
